Major refactor to create independent app builds for android and iOS.
