+++ app/controllers/app_listing_module.rb
@@ -1,8 +1,18 @@
+require 'ios_build'
+
 module AppListingModule
 
-  def app_builds(app_name,app_dir)
+  IOS_APP_TYPE = 1
+  ANDROID_APP_TYPE = 2
+
+  def app_builds(app_name,app_dir,app_type)
     app_builds = subdirectories(app_dir.join(app_name)).map do |build_version|
-      AppBuild.new(app_dir, app_name, build_version)
+      if app_type == IOS_APP_TYPE
+        IosBuild.new(app_dir, app_name, build_version)
+      elsif app_type == ANDROID_APP_TYPE
+        AppBuild.new(app_dir, app_name, build_version)
+      end
+        
     end
     app_builds.sort.reverse
   end
